Move common file package in cmd to internal
