+++ cmd/bee-file/main.go
@@ -14,7 +14,7 @@ import (
 	"os"
 	"path/filepath"
 
-	cmdfile "github.com/ethersphere/bee/cmd/file"
+	cmdfile "github.com/ethersphere/bee/cmd/internal/file"
 	"github.com/ethersphere/bee/pkg/collection/entry"
 	"github.com/ethersphere/bee/pkg/file/joiner"
 	"github.com/ethersphere/bee/pkg/file/splitter"
