+++ cmd/internal/file/io_test.go
@@ -16,7 +16,7 @@ import (
 	"strconv"
 	"testing"
 
-	cmdfile "github.com/ethersphere/bee/cmd/file"
+	cmdfile "github.com/ethersphere/bee/cmd/internal/file"
 	"github.com/ethersphere/bee/pkg/api"
 	"github.com/ethersphere/bee/pkg/file"
 	"github.com/ethersphere/bee/pkg/logging"
